backend/requirements.txt
Django==6.0.4
PyJWT==2.12.1
asgiref==3.11.1
django-cors-headers==4.9.0
django-ninja==1.6.2
django-ninja-jwt==5.4.4
openai==2.31.0
psycopg==3.3.3
pydantic==2.12.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2

# Dev / test
pytest==9.0.3
pytest-django==4.12.0
ruff==0.15.10

